requirements.txt
fastapi
uvicorn
pillow
inference-sdk
python-multipart